Correct syntax error and ensure they can't screw things up again
